Create String-br
Hi, I confess that my English is terrible, it is not a mistake that you wanted to tell me. But I'm putting the information from the String-br.lua file that I created. What would be the next step? or should I do it in another way?

Thank you for your patience